initial mysql support

+++ chef/cookbooks/mysql/recipes/server.rb
@@ -70,7 +70,7 @@ service "mysql" do
     start_command "start mysql"
   end
   supports :status => true, :restart => true, :reload => true
-  action :nothing
+  action :enable
 end
 
 link value_for_platform([ "centos", "redhat", "suse" , "fedora" ] => {"default" => "/etc/my.cnf"}, "default" => "/etc/mysql/my.cnf") do
@@ -156,21 +156,22 @@ template "/etc/mysql/conf.d/emergency_init_file" do
   action :create
 end
 
-
-script "fix_perms_hack" do
-  interpreter "bash"
-  user "root"
-  cwd "/tmp"
-  code <<-EOH
-  /etc/init.d/mysql stop
-  chmod 644 /etc/mysql/conf.d/emergency_init_file
-  /usr/bin/mysqld_safe --init-file=/etc/mysql/conf.d/emergency_init_file &
-  sleep 10
-  killall mysqld
-  chmod 600 /etc/mysql/conf.d/emergency_init_file
-  /etc/init.d/mysql start
-  EOH
-  not_if "/usr/bin/mysql -u root #{node['mysql']['server_root_password'].empty? ? '' : '-p' }#{node['mysql']['server_root_password']} -e 'show databases;'"
+if not platform?(%w{centos redhat})
+  script "fix_perms_hack" do
+    interpreter "bash"
+    user "root"
+    cwd "/tmp"
+    code <<-EOH
+    /etc/init.d/mysql stop
+    chmod 644 /etc/mysql/conf.d/emergency_init_file
+    /usr/bin/mysqld_safe --init-file=/etc/mysql/conf.d/emergency_init_file &
+    sleep 10
+    killall mysqld
+    chmod 600 /etc/mysql/conf.d/emergency_init_file
+    /etc/init.d/mysql start
+    EOH
+    not_if "/usr/bin/mysql -u root #{node['mysql']['server_root_password'].empty? ? '' : '-p' }#{node['mysql']['server_root_password']} -e 'show databases;'"
+  end
 end
 
 # End hackness
